#89 Refatorando APIs de manipulaçao do user
